food_calorie도 floatfield로 수정

+++ diets/models.py
@@ -38,7 +38,7 @@ class DietList(CommonModel):
 
 class SelectedDiet(CommonModel):
     food_name = models.CharField(max_length=250)
-    food_calorie = models.PositiveIntegerField()
+    food_calorie = models.FloatField()
     food_gram = models.PositiveIntegerField()
 
     def __str__(self) -> str:
